base-files: Create /run/lock directory
When starting an image openwrt-image-full, the services (e.g. netifd)
are not started because the '/run/lock' directory does not exist.

This patch creates directory '/run/lock' in the rootfs.

We are not sure if this is the right solution. We observed that the
'/etc/fstab' contains a line to mount '/run' in tempfs. But this does
not happen. Which mechanism would normally create '/run/lock' at
startup?

Signed-off-by: Kevin Zhang <kevin.zhang3@cn.bosch.com>
Signed-off-by: Mark Jonas <mark.jonas@de.bosch.com>